feat: add workflow nuxt example
also added nuxt serve method. While adding serve for nuxt, had to change the TResponse from extending Response because in nuxt I had to get the result (which is the workflow id) as string so that I could return a status/body object.

Also moved SveltekitServeParameters to types as WorkflowServeParametersWithClient for other potential frameworks.