feat: add error handling
there are two new parameters in Workflow options: failureFunction and failureUrl. Only one should be passed in any case.

If failureFunction is passed, we pass the workflow URL as upstash-failure-callback. When upstash calls us, we check upstash-workflow-is-failure header and if it is set, we trigger failureFunction.

If failureUrl is set, we simply pass it as upstash-failure-callback and it is called if all retries are exhausted.