fix: update platform handlers to return Response with body when status: 500
when we simply raised an error, our handlers returned a response with status: 500 but with an empty body. The message would be in statusText field which is not processed by QStash. This PR fixes the issue by putting the serve execution in platform handlers in a try catch. This way, we return a custom body typed as FailureFunctionPayload.

Additionally, fromSafePromise is removed from triggerFirstInvocation because it gets stuck when there is an error in the promise in it.

Finally, the TResponse generic type is changed to extend Response. This was previously removed because I thought I couldn't return Response in nuxt. Which is incorrect.