fix: resolve qstashClient & receiver parameter issues
When used in a package, I was unable to pass new Client to qstashClient getting this error:
Type 'import(qstash-js/dist/client-BBvt1wQq).f' is not assignable to type 'import(qstash-js/dist/client-BBvt1wQq).f'. Two different types with this name exist, but they are unrelated.
  Types have separate declarations of a private property 'token'.ts(2719)

this change fixes the error by adding interfaces for Client and Receiver.