Fix resend integration headers (#219)
* fix: resend integration headers

content-type header wasn't being sent, so resend couldn't process the request. But the integration worked before. So I am suspecting that QStash used to set a default content header but it doesn't anymore. To fix the issue, I added a utility to properly merge the headers and added tests to make sure that we don't miss this behavior.

* fix: add method option to providers