fix: deprecate verifySignatureEdge from nextjs
because of NextFetchEvent parameter, we were getting 'Type NextFetchEvent is not a valid type for the function's second argument' after nextjs version 15. Initially, we were going to remove the parameter. https://github.com/upstash/qstash-js/pull/228 But this would break the code of users who are still on next 13/14 etc, use verifySignatureEdge and the NextFetchEvent parameter.

To avoid breaking code of users on older nextjs version while asking them to remove verifySignatureEdge, we deprecate verifySignatureEdge. verifySignatureAppRouter can be used instead.